shift types around, make overseer half aware
