+++ roadmap/implementers-guide/src/types/overseer-protocol.md
@@ -54,8 +54,8 @@ enum ApprovalVotingMessage {
 	/// Check if the assignment is valid and can be accepted by our view of the protocol.
 	/// Should not be sent unless the block hash is known.
 	CheckAndImportAssignment(
-		Hash, 
-		AssignmentCert, 
+		Hash,
+		AssignmentCert,
 		ValidatorIndex,
 		ResponseChannel<VoteCheckResult>,
 	),
@@ -68,14 +68,15 @@ enum ApprovalVotingMessage {
 		ResponseChannel<VoteCheckResult>,
 	),
 	/// Returns the highest possible ancestor hash of the provided block hash which is
-	/// acceptable to vote on finality for. 
+	/// acceptable to vote on finality for.
 	/// The `BlockNumber` provided is the number of the block's ancestor which is the
 	/// earliest possible vote.
-	/// 
-	/// It can also return the same block hash, if that is acceptable to vote upon. 
+	///
+	/// It can also return the same block hash, if that is acceptable to vote upon.
 	/// Return `None` if the input hash is unrecognized.
 	ApprovedAncestor(Hash, BlockNumber, ResponseChannel<Option<Hash>>),
 }
+
 ```
 
 ## Approval Networking
@@ -122,8 +123,8 @@ Messages received by the availability recovery subsystem.
 enum AvailabilityRecoveryMessage {
 	/// Recover available data from validators on the network.
 	RecoverAvailableData(
-		CandidateDescriptor, 
-		SessionIndex, 
+		CandidateDescriptor,
+		SessionIndex,
 		ResponseChannel<Option<AvailableData>>,
 	),
 }
@@ -409,7 +410,7 @@ enum RuntimeApiRequest {
 	SessionIndex(ResponseChannel<SessionIndex>),
 	/// Get the validation code for a specific para, using the given occupied core assumption.
 	ValidationCode(ParaId, OccupiedCoreAssumption, ResponseChannel<Option<ValidationCode>>),
-	/// Fetch the historical validation code used by a para for candidates executed in 
+	/// Fetch the historical validation code used by a para for candidates executed in
 	/// the context of a given block height in the current chain.
 	HistoricalValidationCode(ParaId, BlockNumber, ResponseChannel<Option<ValidationCode>>),
 	/// with the given occupied core assumption.
